+++ src/bookshelf/reading-journey/position/page.ts
@@ -1,10 +1,19 @@
 import { Part, Position } from './position'
+import { Book } from '../../book/book'
 
 export class Page implements Position {
     constructor(private value: number) {}
 
     public first(): Position {
         return new Page(1)
+    }
+
+    public last(book: Book): Position | null {
+        if (book.metadata.pages === undefined) {
+            return null
+        }
+
+        return new Page(book.metadata.pages)
     }
 
     public next(): Position {